+++ crates/ruma-client-api/CHANGELOG.md
@@ -3,6 +3,11 @@
 Breaking changes:
 
 - Upgrade `js_option` to v0.2.0
+- Make the `ErrorBody::Standard` variant a newtype around `StandardErrorBody`.
+- The `message` field of `StandardErrorBody` is an `Option<String>`. It was
+  never required in the spec.
+- `StandardErrorBody` is non-exhaustive. Allowing to add fields in the future
+  without it being a breaking change.
 
 Fixes: